Camera: Check for minimum depth Jpeg output size
The Jpeg encoder assumes a miminum output buffer size
which should at least be sufficient for the headers.
In worst case scenario, the Jpeg buffer can be overwritten
before the library detects and flags the error.
Check the depth Jpeg output buffer size before trying
to encode the final image and re-use the minimum value
as defined in the Camera3Device.

FLAG: EXEMPT bugfix
Bug: 358210775
Test: atest frameworks/av/services/camera/libcameraservice/tests/DepthProcessorTest.cpp
Change-Id: I93021e87681864da76f447017dda00981661f9f0